protocol ink lib done (#59)
