+++ contracts/mixer/merkle_tree.rs
@@ -7,6 +7,7 @@ use ink_storage::Mapping;
 use poseidon::PoseidonRef;
 
 use ink_prelude::vec;
+use protocol_ink_lib::zeroes::zeroes;
 
 #[derive(Default, Debug, SpreadLayout, SpreadAllocate)]
 #[cfg_attr(feature = "std", derive(StorageLayout))]
@@ -44,7 +45,7 @@ impl MerkleTree {
         for i in 0..self.levels {
             if current_index % 2 == 0 {
                 left = current_level_hash;
-                right = zeroes::zeroes(i);
+                right = zeroes(i);
                 self.filled_subtrees.insert(i, &current_level_hash);
             } else {
                 left = self.filled_subtrees.get(&i).unwrap_or_default();
