--- contracts/vanchor/zeroes.rs
@@ -1,133 +1,0 @@
-pub fn zeroes(index: u32) -> [u8; 32] {
-    match index {
-        0 => [
-            108, 175, 153, 72, 237, 133, 150, 36, 226, 65, 231, 118, 15, 52, 27, 130, 180, 93, 161,
-            235, 182, 53, 58, 52, 243, 171, 172, 211, 96, 76, 229, 47,
-        ],
-        1 => [
-            248, 34, 66, 56, 136, 142, 122, 46, 31, 162, 21, 166, 176, 107, 140, 120, 153, 177,
-            194, 7, 134, 120, 193, 204, 120, 108, 184, 108, 45, 127, 227, 19,
-        ],
-        2 => [
-            101, 39, 98, 220, 50, 158, 26, 164, 39, 239, 223, 94, 246, 12, 213, 58, 214, 143, 236,
-            62, 128, 194, 232, 150, 104, 50, 44, 53, 250, 38, 113, 33,
-        ],
-        3 => [
-            238, 126, 57, 215, 172, 76, 243, 93, 238, 136, 131, 97, 12, 35, 109, 77, 178, 225, 24,
-            218, 58, 158, 90, 125, 0, 145, 62, 155, 26, 166, 40, 14,
-        ],
-        4 => [
-            135, 53, 1, 165, 207, 165, 10, 47, 249, 54, 188, 233, 146, 226, 176, 173, 95, 209, 94,
-            66, 197, 173, 107, 83, 57, 152, 151, 166, 71, 52, 149, 39,
-        ],
-        5 => [
-            252, 175, 227, 224, 204, 194, 109, 223, 194, 117, 113, 196, 1, 120, 12, 65, 149, 48,
-            243, 21, 211, 175, 167, 246, 16, 29, 185, 191, 237, 145, 65, 25,
-        ],
-        6 => [
-            31, 116, 71, 90, 132, 214, 233, 29, 115, 204, 192, 159, 7, 199, 61, 100, 250, 54, 121,
-            146, 241, 186, 109, 81, 144, 17, 215, 23, 206, 222, 51, 23,
-        ],
-        7 => [
-            25, 146, 153, 239, 178, 210, 33, 66, 232, 176, 77, 159, 225, 90, 191, 114, 165, 122,
-            10, 93, 249, 23, 29, 216, 57, 219, 117, 220, 167, 85, 120, 38,
-        ],
-        8 => [
-            13, 98, 39, 124, 2, 65, 54, 14, 155, 62, 227, 39, 83, 102, 186, 238, 133, 9, 204, 14,
-            52, 138, 35, 216, 106, 195, 180, 54, 24, 225, 132, 17,
-        ],
-        9 => [
-            206, 197, 214, 19, 34, 123, 26, 56, 21, 114, 106, 62, 43, 29, 252, 216, 171, 11, 169,
-            250, 27, 171, 80, 83, 245, 215, 53, 161, 131, 171, 2, 7,
-        ],
-        10 => [
-            215, 106, 194, 174, 254, 207, 145, 50, 127, 164, 247, 174, 33, 102, 165, 29, 13, 227,
-            178, 171, 43, 136, 206, 87, 140, 253, 76, 129, 222, 192, 236, 46,
-        ],
-        11 => [
-            71, 59, 211, 98, 196, 29, 98, 25, 83, 0, 206, 48, 221, 87, 81, 5, 1, 133, 176, 199,
-            182, 133, 53, 131, 213, 85, 193, 69, 33, 192, 11, 40,
-        ],
-        12 => [
-            91, 252, 90, 13, 15, 198, 185, 114, 194, 46, 203, 44, 115, 183, 137, 33, 30, 236, 203,
-            210, 172, 216, 174, 244, 167, 160, 31, 29, 34, 50, 81, 4,
-        ],
-        13 => [
-            102, 211, 113, 206, 113, 67, 60, 161, 0, 30, 207, 136, 77, 172, 214, 59, 165, 107, 218,
-            78, 190, 171, 112, 18, 75, 164, 88, 191, 204, 39, 244, 39,
-        ],
-        14 => [
-            239, 44, 11, 247, 13, 236, 220, 213, 166, 28, 223, 101, 27, 231, 253, 215, 222, 191,
-            146, 174, 147, 100, 138, 143, 110, 242, 100, 80, 174, 234, 23, 22,
-        ],
-        15 => [
-            202, 171, 117, 165, 206, 234, 111, 121, 194, 206, 102, 218, 200, 165, 123, 173, 99,
-            238, 49, 28, 195, 3, 87, 67, 21, 27, 234, 208, 0, 180, 198, 32,
-        ],
-        16 => [
-            236, 94, 100, 101, 41, 115, 24, 54, 167, 137, 56, 206, 254, 232, 103, 237, 248, 197,
-            167, 173, 189, 87, 142, 58, 245, 35, 135, 67, 219, 157, 88, 9,
-        ],
-        17 => [
-            155, 7, 54, 50, 191, 102, 17, 220, 37, 84, 92, 53, 253, 91, 211, 203, 14, 79, 57, 243,
-            32, 178, 183, 125, 83, 255, 165, 56, 167, 182, 100, 0,
-        ],
-        18 => [
-            159, 158, 124, 85, 94, 121, 103, 45, 173, 132, 247, 93, 188, 29, 232, 94, 13, 121, 255,
-            116, 53, 124, 137, 94, 5, 213, 177, 129, 98, 229, 93, 9,
-        ],
-        19 => [
-            99, 155, 90, 133, 223, 235, 190, 224, 111, 61, 205, 78, 241, 96, 76, 77, 254, 62, 24,
-            137, 66, 193, 110, 58, 150, 33, 170, 135, 40, 46, 207, 17,
-        ],
-        20 => [
-            187, 188, 119, 234, 147, 19, 87, 242, 72, 154, 76, 196, 174, 46, 77, 55, 199, 132, 30,
-            14, 124, 98, 115, 111, 91, 166, 159, 23, 192, 111, 15, 43,
-        ],
-        21 => [
-            121, 50, 177, 96, 164, 62, 96, 241, 44, 123, 83, 61, 39, 106, 251, 70, 140, 37, 100,
-            253, 242, 219, 136, 249, 192, 249, 171, 194, 55, 182, 253, 22,
-        ],
-        22 => [
-            255, 240, 220, 78, 69, 26, 219, 206, 247, 63, 112, 255, 125, 228, 102, 202, 231, 18,
-            204, 249, 109, 55, 76, 173, 77, 18, 246, 68, 233, 215, 187, 33,
-        ],
-        23 => [
-            98, 153, 77, 69, 211, 176, 78, 232, 182, 217, 122, 162, 146, 59, 91, 98, 161, 170, 123,
-            19, 15, 89, 143, 70, 62, 150, 28, 11, 34, 248, 132, 39,
-        ],
-        24 => [
-            16, 252, 9, 229, 87, 186, 152, 203, 117, 194, 217, 5, 169, 76, 167, 167, 214, 179, 16,
-            216, 170, 193, 140, 47, 20, 52, 117, 91, 166, 225, 172, 22,
-        ],
-        25 => [
-            19, 160, 150, 42, 25, 246, 56, 54, 94, 166, 37, 167, 94, 102, 240, 41, 63, 13, 225,
-            143, 189, 222, 36, 81, 38, 36, 140, 106, 140, 6, 40, 35,
-        ],
-        26 => [
-            20, 172, 201, 177, 202, 74, 245, 173, 0, 111, 115, 32, 193, 81, 231, 229, 67, 32, 194,
-            210, 196, 180, 17, 36, 2, 40, 240, 225, 27, 153, 219, 45,
-        ],
-        27 => [
-            192, 156, 32, 204, 56, 77, 89, 165, 24, 229, 47, 95, 21, 192, 119, 131, 181, 30, 82,
-            112, 79, 70, 106, 5, 149, 235, 234, 16, 132, 121, 19, 1,
-        ],
-        28 => [
-            80, 77, 245, 152, 55, 106, 253, 207, 36, 44, 76, 2, 33, 87, 244, 122, 90, 5, 212, 181,
-            15, 248, 30, 208, 208, 240, 38, 21, 230, 26, 45, 32,
-        ],
-        29 => [
-            135, 241, 176, 96, 154, 4, 171, 10, 48, 150, 167, 150, 64, 111, 189, 190, 245, 34, 80,
-            97, 121, 95, 118, 242, 41, 129, 116, 83, 52, 50, 171, 35,
-        ],
-        30 => [
-            160, 138, 218, 95, 86, 180, 30, 11, 21, 87, 76, 148, 219, 172, 9, 169, 157, 121, 22,
-            135, 145, 189, 248, 188, 120, 227, 71, 137, 95, 88, 21, 31,
-        ],
-        // Reuse index 0 for wildcard
-        _ => [
-            108, 175, 153, 72, 237, 133, 150, 36, 226, 65, 231, 118, 15, 52, 27, 130, 180, 93, 161,
-            235, 182, 53, 58, 52, 243, 171, 172, 211, 96, 76, 229, 47,
-        ],
-    }
-}
