+++ protocol-ink-lib/src/zeroes.rs
